Changes in chat UI and added two services of AI in supabase.
