+++ my-portfolio-angular/portfolio-site/src/app/layouts/main-layout/main-layout.component.html
@@ -10,7 +10,7 @@
     </main>
     <app-footer></app-footer>
   </div>
-  <app-chat-bot [position]="'bottom-right'" [title]="'Portfolio Assistant'" 
+  <app-chat-bot [position]="'bottom-right'" [title]="'AiNg Portfolio Assistant'" 
     [placeholder]="'Ask about my projects...'">
   </app-chat-bot>
 </div>
