+++ my-portfolio-angular/portfolio-site/src/app/shared/components/chat-bot/chat-bot.component.html
@@ -12,11 +12,8 @@
     </div>
     
     <div #messagesContainer class="chat-messages">
-      <div *ngFor="let msg of messages" 
-           [class.user]="msg.sender === 'user'"
-           [class.loading]="msg?.loading"
-           [class.error]="msg.error"
-           class="message">
+      <div *ngFor="let msg of messages" [class.user]="msg.sender === 'user'" [class.loading]="msg?.loading"
+           [class.error]="msg.error" class="message">
         
         <div class="message-content">
           <!-- Loading indicator -->
@@ -37,22 +34,17 @@
           </div>
         </div>
         
-        <div class="message-time" *ngIf="!msg.loading">
+        <div class="message-time"  *ngIf="!msg.loading">
           {{msg.timestamp | date:'shortTime'}}
         </div>
       </div>
     </div>
     
     <div class="chat-input">
-      <input type="text" 
-             [(ngModel)]="newMessage" 
-             [placeholder]="placeholder" 
-             [disabled]="isSending"
-             (keyup.enter)="sendMessage()">
-      
+      <input type="text" [(ngModel)]="newMessage" [placeholder]="placeholder" [disabled]="isSending" (keyup.enter)="sendMessage()">
       <button (click)="sendMessage()" [disabled]="isSending || !newMessage.trim()">
-        <i *ngIf="!isSending" class="fas fa-paper-plane"></i>
-        <i *ngIf="isSending" class="fas fa-spinner fa-spin"></i>
+        <i class="fas fa-paper-plane"></i>
+        <!-- <i *ngIf="isSending" class="fas fa-spinner fa-spin"></i> -->
       </button>
     </div>
   </div>
