+++ my-portfolio-angular/portfolio-site/src/app/shared/components/chat-bot/chat-bot.component.scss
@@ -67,7 +67,7 @@
 
       .message {
         margin-bottom: 15px;
-        max-width: 80%;
+        max-width: 95%;
 
         .message-content {
           padding: 10px 15px;
@@ -141,6 +141,42 @@
           background: var(--global-secondary-color);
         }
       }
+    }
+  }
+
+  .typing-indicator {
+    display: flex;
+    align-items: center;
+    height: 22px;
+    margin: 4px 0 0 8px;
+    gap: 4px;
+
+    span {
+      display: inline-block;
+      width: 8px;
+      height: 8px;
+      background: var(--global-primary-color);
+      border-radius: 50%;
+      opacity: 0.6;
+      animation: typing-bounce 1.2s infinite both;
+    }
+
+    span:nth-child(2) {
+      animation-delay: 0.2s;
+    }
+    span:nth-child(3) {
+      animation-delay: 0.4s;
+    }
+  }
+
+  @keyframes typing-bounce {
+    0%, 80%, 100% {
+      transform: translateY(0);
+      opacity: 0.6;
+    }
+    40% {
+      transform: translateY(-8px);
+      opacity: 1;
     }
   }
 }
